Revert "test: Add SynapseInternal E2E tests (#1683)"
This reverts commit 59b900c6fa8225f7708ed2a851500c0ff6e81132.
