--- core/src/test/scala/com/microsoft/azure/synapse/ml/nbtest/SynapseExtension/SynapseExtensionSchemas.scala
@@ -1,16 +1,0 @@
-// Copyright (C) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in project root for information.
-
-package com.microsoft.azure.synapse.ml.nbtest.SynapseExtension
-
-case class Artifact (
-  objectId: String,
-  displayName: String
-)
-
-case class SparkJobDefinitionExecutionResponse (
-  statusString: String,
-  artifactJobInstanceId: String,
-  serviceExceptionJson: Option[String]
-)
-
